+++ src/pytorch_ie/taskmodules/transformer_re_text_classification.py
@@ -556,7 +556,7 @@ class TransformerRETextClassificationTaskModule(TaskModuleType, ChangesTokenizer
         logits = model_output["logits"]
 
         output_label_probs = logits.sigmoid() if self.multi_label else logits.softmax(dim=-1)
-        output_label_probs = output_label_probs.detach().cpu().numpy()
+        output_label_probs = output_label_probs.detach().cpu().float().numpy()
 
         unbatched_output = []
         if self.multi_label:
